Fix DockerContainer#add_exposed_ports.
Avoid overriding the PortBinding settings set by add_fixed_exposed_port
when the same port is exposed with add_exposed_ports in a later stage.

This allows to the modules to call to a default add_exposed_ports on
start without overriding the previous port bindings set by the user on
purpose.